camera/requirements.txt
fastapi==0.61.1
xlrd==1.2.0
requests==2.24.0
uvicorn
fastapi-utils==0.2.1
mongoengine==0.20.0
PyJWT==1.7.1
python-multipart
python-jose==3.2.0
apscheduler==3.7.0
werkzeug
aiofiles==0.6.0
cryptocode==0.1